+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## [Unreleased]
+- Add routine to delete branches
 - Add routine to fetch meta information
 - add routine to get local and remote branches
 - add repository check